src/requirements.txt
PyJWT
python-dotenv
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Bcrypt
Flask-Migrate
Flask-Cors